% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.R
\docType{data}
\name{co}
\alias{co}
\title{Tibble con i dati di co per 2 stazioni della regione MOLISE}
\format{
Un tibble con 8 colonne e 1462 osservazioni
}
\usage{
co
}
\description{
Tibble con i dati di co per 2 stazioni della regione MOLISE
}
\keyword{datasets}
